Added / updated files to for password functions. These will work with new access plugins (adauth - specifically).
